Updated code for installementAmount calculation

+++ src/main/java/com/paywallet/userservice/user/model/wrapperAPI/DepositAllocationResponseWrapperModel.java
@@ -10,7 +10,7 @@ public class DepositAllocationResponseWrapperModel {
 
 	private String emailId;
 	private String cellPhone;
-	private Integer installmentAmount;
+	private double installmentAmount;
 	private Integer numberOfInstallments;
 	private String virtualAccountNumber;
 	@JsonInclude(value = Include.NON_NULL)
